# src/GovUK.Dfe.ClamAV.Api.Client/Generated/swagger.json
{
  "x-generator": "NSwag v14.6.2.0 (NJsonSchema v11.5.2.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "ClamAv Api",
    "version": "1.0.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "tags": [
          "Health"
        ],
        "operationId": "HealthCheck",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/version": {
      "get": {
        "tags": [
          "Health"
        ],
        "operationId": "GetVersion",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/scan": {
      "post": {
        "tags": [
          "Scan"
        ],
        "operationId": "ScanSync",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/scan/async": {
      "post": {
        "tags": [
          "Scan"
        ],
        "operationId": "ScanAsync",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        }
      }
    },
    "/scan/async/url": {
      "post": {
        "tags": [
          "Scan"
        ],
        "operationId": "ScanAsyncUrl",
        "requestBody": {
          "x-name": "urlRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanUrlRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "202": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        }
      }
    },
    "/scan/async/{jobId}": {
      "get": {
        "tags": [
          "Scan"
        ],
        "operationId": "GetScanStatus",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        }
      }
    },
    "/scan/jobs": {
      "get": {
        "tags": [
          "Scan"
        ],
        "operationId": "ListJobs",
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ScanUrlRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "url": {
            "type": "string"
          },
          "isBase64": {
            "type": "boolean"
          }
        }
      }
    }
  }
}